reduce amount of generated xml parsing logic
